requirements.txt
streamlit
pptx
PyPDF2
pandas
transformers
torch
scikit-learn
python-docx
